Update LIO Model

+++ core/simulator/src/main/java/it/unicam/quasylab/sibilla/core/models/lio/LIOCountingState.java
@@ -70,7 +70,7 @@ public class LIOCountingState implements LIOState<LIOCountingState> {
     }
 
     @Override
-    public int size() {
+    public double size() {
         return size;
     }
